fix: reflect home UI change
