--- src/components/Meeting/style.module.scss
@@ -1,104 +1,0 @@
-@import 'src/styles/_color.scss';
-@import 'src/styles/_variables.scss';
-@import 'src/styles/_mixin.scss';
-
-.meeting {
-  display: flex;
-  min-height: 65px;
-  padding: 15px;
-  align-items: center;
-  border-bottom: 1px solid $gray-250;
-  background-color: $white;
-  cursor: pointer;
-  &:last-child {
-    border-bottom: none;
-  }
-
-  .meetingInfo {
-    width: 50%;
-    margin: auto 0;
-    margin-right: 20px;
-    font-size: 15px;
-
-    .title {
-      @include textOverFlow;
-      display: block;
-      margin-bottom: 10px;
-      font-weight: 700;
-    }
-    .college,
-    .num {
-      display: block;
-      font-size: 14px;
-      font-weight: 300;
-    }
-    .college {
-      @include textOverFlow;
-      margin-bottom: 5px;
-      font-weight: 400;
-    }
-  }
-
-  .imageList {
-    width: calc(50% - 20px);
-    @include mobile {
-      div {
-        height: $boardMobileImage;
-        .profileImg > img {
-          width: $boardMobileImage;
-          height: $boardMobileImage;
-        }
-        .isClicked ~ .profileImg {
-          margin-left: 15px;
-        }
-      }
-    }
-    @media (min-width: #{$breakpoint-tablet}) {
-      div {
-        height: $boardDesktopImage;
-        .profileImg > img {
-          width: $boardDesktopImage;
-          height: $boardDesktopImage;
-        }
-        .profileBtn:hover ~ .profileImg {
-          margin-left: 15px;
-        }
-      }
-    }
-
-    div {
-      position: relative;
-      display: flex;
-      flex-direction: row-reverse;
-      .profileImg {
-        display: inline-block;
-        margin-left: -20px;
-        img {
-          border-radius: 20%;
-        }
-        .alone {
-          width: $boardDesktopImage;
-          height: $boardDesktopImage;
-        }
-        &:last-child {
-          margin: 0;
-        }
-      }
-      .profileBtn {
-        position: absolute;
-        width: $boardMobileImage / 2;
-        height: $boardMobileImage / 2;
-        bottom: 0;
-        z-index: 5;
-        line-height: 1.6;
-        border-radius: 20%;
-        background-color: $moreImgBtn;
-        text-align: center;
-        font-size: 17px;
-        font-weight: 300;
-        color: $white;
-        cursor: pointer;
-      }
-    }
-  }
-}
